Reorg complete
