+++ makefiles/config_s2740vc_bootloader.mk
@@ -7,8 +7,6 @@
 #
 MODULES		+= drivers/boards/s2740vc/bootloader
 
-SRC_SEARCH = 	can src common uavcan
-
-INCLUDE_DIRS += $(addprefix $(PX4_MODULE_SRC)$(MODULES)/,$(SRC_SEARCH))
+INCLUDE_DIRS += $(PX4_MODULE_SRC)$(MODULES) $(PX4_BOOTLOADER_BASE)include $(PX4_MODULE_SRC)/modules/systemlib
 
 START_UP_FILES = stm32_vectors.c
