+++ src/drivers/boards/px4cannode-v1/bootloader/boot_config.h
@@ -32,11 +32,12 @@
  ****************************************************************************/
 
 /*
- * @file board_config.h
+ * @file boot_config.h
  *
- * PX4CANNODEv1 for the bootloader internal definitions
- * This file is related to the parrent folder version but defines
- * differnet usages of the hardware for bootloading
+ * bootloader definitions that configures the behavior and options
+ * of the Boot loader
+ * This file is relies on the parent folder's boot_config.h file and defines
+ * different usages of the hardware for bootloading
  */
 
 #pragma once
@@ -50,8 +51,8 @@
  * todo:make this be pulled in from a targed's build
  * files in nuttx*/
 
-#include "../../board_config.h"
-#include "protocol.h"
+#include "../board_config.h"
+#include "uavcan.h"
 #include <nuttx/compiler.h>
 #include <stdint.h>
 
@@ -64,11 +65,11 @@
 #define OPT_NODE_INFO_RATE_MS   200
 #define OPT_BL_NUMBER_TIMERS    6
 
-#define OPT_WAIT_FOR_GETNODEINFO                    1
-/* #define OPT_WAIT_FOR_GETNODEINFO_JUMPER_GPIO         */
-/* #define OPT_WAIT_FOR_GETNODEINFO_JUMPER_GPIO_INVERT 1 */
+#define OPT_WAIT_FOR_GETNODEINFO                    0
+#define OPT_WAIT_FOR_GETNODEINFO_JUMPER_GPIO        1
+#define OPT_WAIT_FOR_GETNODEINFO_JUMPER_GPIO_INVERT 1
 
-/* #define OPT_ENABLE_WD */
+#define OPT_ENABLE_WD           1
 
 #define OPT_RESTART_TIMEOUT_MS  20000u
 
@@ -105,97 +106,3 @@
  */
 #define GPIO_GETNODEINFO_JUMPER (BUTTON_BOOT0n&~(GPIO_EXTI))
 
-
-/****************************************************************************
- * Public Type Definitions
- ****************************************************************************/
-
-typedef enum {
-
-    reset,
-    autobaud_start,
-    autobaud_end,
-    allocation_start,
-    allocation_end,
-    fw_update_start,
-    fw_update_erase_fail,
-    fw_update_invalid_response,
-    fw_update_timeout,
-    fw_update_invalid_crc,
-    jump_to_app,
-} uiindication_t;
-/************************************************************************************
- * Public data
- ************************************************************************************/
-
-#ifndef __ASSEMBLY__
-
-/****************************************************************************
- * Public Function Prototypes
- ****************************************************************************/
-
-/************************************************************************************
- * Name: stm32_boarddeinitialize
- *
- * Description:
- *   This function is called by the bootloader code priore to booting
- *   the application. Is should place the HW into an benign initialized state.
- *
- ************************************************************************************/
-
-void stm32_boarddeinitialize(void);
-
-/****************************************************************************
- * Name: board_get_product_name
- *
- * Description:
- *   Called to retrive the product name. The retuned alue is a assumed
- *   to be written to a pascal style string that will be length prefixed
- *   and not null terminated
- *
- * Input Parameters:
- *    product_name - A pointer to a buffer to write the name.
- *    maxlen       - The imum number of chatater that can be written
- *
- * Returned Value:
- *   The length of charaacters written to the buffer.
- *
- ****************************************************************************/
-
-uint8_t board_get_product_name(uint8_t * product_name, size_t maxlen);
-
-/****************************************************************************
- * Name: board_get_hardware_version
- *
- * Description:
- *   Called to retrive the hardware version information.
- *
- * Input Parameters:
- *    hw_version - A pointer to a uavcan_hardwareversion_t.
- *
- * Returned Value:
- *   None
- *
- ****************************************************************************/
-
-void board_get_hardware_version(uavcan_hardwareversion_t * hw_version);
-
-/****************************************************************************
- * Name: board_indicate
- *
- * Description:
- *   Provides User feedback to indicate the state of the bootloader
- *   on board specific  hardware.
- *
- * Input Parameters:
- *    indication - A member of the uiindication_t
- *
- * Returned Value:
- *   None
- *
- ****************************************************************************/
-
-void board_indicate(uiindication_t indication);
-
-
-#endif /* __ASSEMBLY__ */
