+++ src/drivers/bootloaders/include/boot_app_shared.h
